YD-424 Now caches the certificate and key
Along with this corrected the name of the enablement property.
