+++ core/src/main/resources/application.properties
@@ -56,7 +56,7 @@ yona.security.maxCreateUserAttemptsPerTimeWindow = 2
 yona.security.maxUpdateUserAttemptsPerTimeWindow = 2
 yona.security.corsAllowed = true
 
-yona.security.appleMobileConfigSigning.isSigningEnabled = true
+yona.security.appleMobileConfigSigning.signingEnabled = false
 yona.security.appleMobileConfigSigning.signingCertificateFile = smime.crt
 yona.security.appleMobileConfigSigning.signingKeyFile = smime.key
 yona.security.appleMobileConfigSigning.password = Yona1
